Add files via upload
 BUG TRACKING SYSTEM AND SOLUTIONS FOR IMPROVING SOFTWARE RELIABLITY
           Bug Tracking for Improving Software Reliability (BTS) is an automated system that can be useful to employees and the managers in any functional organization. Bug Tracking System gives the facility to define the tasks in the organization and also allows the managers to track the bugs spent by the employee for that particular task. A report generation facility is supported in BTS that allows the managers to analyze which are those skills by employee are utilized and those which are not utilized. This tool can help managers for Bug estimation per project or application. This tool helps employees to document their Bugs and analyze 
This project aims at creation of a Bug Tracking System. This project will be accessible to all developers and its facility allows developers to focus on creating the database schema and while letting the application server define table based on the fields in JSP and relationships between them. This system provides the following facilities.